fix(pdp): fix proving hot-looping, proving early, additional logs (#777)
* fix(pdp): delay proving by one epoch

In logs sometimes we observed that the proving was happening to early,
possible cause is reorgs knocking off a head, causing the message to
land a block early. Resolution: delay proving by one epoch
The task will be re-tried in any case but this will reduce noise.

Signed-off-by: Jakub Sztandera <oss@kubuxu.com>

* chore(pdp): explicity log proving related failures

While these failures get logged through taskhandler, log them explicitly
to get structured logs with `dataSetId`.

Signed-off-by: Jakub Sztandera <oss@kubuxu.com>

* fix(pdp): if we encounter ProvingPeriodNotInitialized, move back to init

Sometimes for a not fully understood reasons (my guess is deletion),
proving period can get de-initialized. This causes NextProvingPeriodTask
to loop in error state over and over. Resolve it by moving from proving
loop to init state.

Signed-off-by: Jakub Sztandera <oss@kubuxu.com>

* fix(pdp): in proving period init state, don't re-try if dataset is empty

We used to trigger PPInit early, before pieces landed. We shifted to
doing it only after pieces land. If we encounter dataset with no pieces
in init state, it should be safe to assume that the leaf count is
accurate.

Signed-off-by: Jakub Sztandera <oss@kubuxu.com>

---------

Signed-off-by: Jakub Sztandera <oss@kubuxu.com>